requirements.txt
fastapi
uvicorn
pytesseract
Pillow
pdf2image
opencv-python